Library Improvements
Add input validation for the `.encode()` methods, and additional unit
testing for the `UnsignedLong` type.

Added a new `utilities` sub-module containing the `hexbytes` helper
method that can be used to format raw `bytes` or `bytearray` values
into hexadecimal encoded strings for review, and the `print_hexbytes`
helper method that can be used to print such strings.

Changed type casting of the decoded return value for the `Bytes` class
when reversed back to a `bytes` type, fixing a previous issue where the
reversed return type was a `list` type.

The `BytesView` class `split` length can now be equal to the length of
the provided `data` value, previously the `split` length had to be less
than the length of the `data` value.

The type hints for the `BytesView` class' `__iter__` and `__next__`
methods have been corrected from `bytesview` to `BytesView`, and
`bytes` to `bytearray` respectively.