Utility Functions
The `Bytes.encode()` and the `Bytes.decode()` methods now support
reversing the order of the provided bytes via the new `reverse` boolean
keyword argument; by default bytes are held and encoded in the order in
which they are provided, regardless of the (byte) `order` keyword
argument.

Added the new `isinstantiable()` utility function, and documentation
and unit tests for all of the utility functions, including `hexbytes`
and `print_hexbytes`.